Lazy create db file
If the db file doesn't exist and we haven't tried to insert anything,
there's no reason to create the file (when you initialize low with a
file that doesn't exist or an empty file it acts exactly the same). We
save IO to when it's actually needed.
This is a partial resolve to typicode/lowdb#42