Improve efficiency of initial parse
In the current code when data was not empty, we would convert the buffer
returned from disk.read to string, twice. The first time was with in the
condition (`data.trim()`) and second was an implicit cast when
JSON.parse is called on data. This can cause performance issues if the
file is big (e.g. a couple of MB).
The new code will convert the buffer to string once. When it arrives to
JSON.parse it's already a string so no need for the implicit casting