fix: pass eslint
There are two types of errors that should not be treated as errors, configure
eslint in the comments to ignore them.

1. @typescript-eslint/require-await, the code does not use await in async, but
we do want a Promise.
2. @typescript-eslint/explicit-module-boundary-types, there is no explicit
return statement in the function, but we really don't care about its return
value.

In order for eslint to make sense, I didn't disable these rules in the file
but to line.